requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
delayed==0.5.1b2
Django==3.0
django-cors-headers==3.2.0
django-jsonfield==1.4.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.1
hiredis==1.0.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mysqlclient==1.4.6
numpy==1.18.1
openapi-codec==1.3.2
packaging==20.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
